Issue 43: Chroma delay (#44)
Chroma delay

+++ ntsc-mtl-demo/Filters/NTSCEffect.swift
@@ -31,7 +31,7 @@ class NTSCEffect {
     var snowAnisotropy: Float
     var chromaPhaseNoiseIntensity: Float16
     var chromaPhaseError: Float16
-    var chromaDelay: (Float, Int)
+    var chromaDelay: (Float16, Int)
     var vhsSettings: VHSSettings
     var chromaVertBlend: Bool
     var chromaLowpassOut: ChromaLowpass
@@ -60,7 +60,7 @@ class NTSCEffect {
         snowAnisotropy: Float = 0.5,
         chromaPhaseNoiseIntensity: Float16 = 0.001,
         chromaPhaseError: Float16 = 0,
-        chromaDelay: (Float, Int) = (0, 0),
+        chromaDelay: (Float16, Int) = (0, 0),
         vhsSettings: VHSSettings = VHSSettings.default,
         chromaVertBlend: Bool = true,
         chromaLowpassOut: ChromaLowpass = ChromaLowpass.full,
